Keep each TabPFN backend with its model and drop torch from core
RDBLearn keeps the v2 and v2.5 recipes it sweeps, TabPFN-Rel the v3 and hosted
recipes, and both call the shared fit_tfm with a spec instead of a name wrapper.
core.tfm imports torch only inside default_device.
